+++ examples/connection.rs
@@ -36,27 +36,18 @@ fn main() {
     info!("[{}] state: {:?}", line!(), conn.status().state());
     info!("[{}] declared queue: {:?}", line!(), queue);
 
-    channel_a
+    let queue = channel_a
         .confirm_select(ConfirmSelectOptions::default())
         .wait()
         .expect("confirm_select");
     info!("[{}] state: {:?}", line!(), conn.status().state());
-
-    let queue = channel_b
-        .queue_declare(
-            "hello",
-            QueueDeclareOptions::default(),
-            FieldTable::default(),
-        )
-        .wait()
-        .expect("queue_declare");
-    info!("[{}] state: {:?}", line!(), conn.status().state());
+    info!("Declared queue {:?}", queue);
 
     let chan = channel_b.clone();
     info!("will consume");
     channel_b
         .basic_consume(
-            &queue,
+            "hello",
             "my_consumer",
             BasicConsumeOptions::default(),
             FieldTable::default(),