+++ src/lib.rs
@@ -39,11 +39,11 @@
 //!   let channel_a = conn.create_channel().wait().expect("create_channel");
 //!   let channel_b = conn.create_channel().wait().expect("create_channel");
 //!
-//!   channel_a.queue_declare("hello", QueueDeclareOptions::default(), FieldTable::default()).wait().expect("queue_declare");
-//!   let queue = channel_b.queue_declare("hello", QueueDeclareOptions::default(), FieldTable::default()).wait().expect("queue_declare");
+//!   let queue = channel_a.queue_declare("hello", QueueDeclareOptions::default(), FieldTable::default()).wait().expect("queue_declare");
+//!   info!("Declared queue {:?}", queue);
 //!
 //!   info!("will consume");
-//!   channel_b.clone().basic_consume(&queue, "my_consumer", BasicConsumeOptions::default(), FieldTable::default()).wait().expect("basic_consume").set_delegate(Box::new(Subscriber { channel: channel_b }));
+//!   channel_b.clone().basic_consume("hello", "my_consumer", BasicConsumeOptions::default(), FieldTable::default()).wait().expect("basic_consume").set_delegate(Box::new(Subscriber { channel: channel_b }));
 //!
 //!   let payload = b"Hello world!";
 //!